Assignment 2
1.Finds and loads files for Processing from a directory tree:
Starting at a path specified on the command line, recursively search each folder in the directory tree rooted at the path, for files with names that match a regular expression. Each matching file is loaded and processed, using the facilities developed in Project #1.
2.Records file dependencies for each processed file:
Dependencies will be based on #include statements3 found in each file. Dependencies are recorded with web links in the constructed web pages. Only #include statements that cite files in the directory tree are recorded with web links.
3.Provides facilities for optionally hiding comments and/or function bodies:
The intent is to allow users some flexibility in what is visible. We want to allow users to:
show or hide comments
show or hide function bodies
show just class names by hiding the entire class scope
This can be done by wrapping text that may be hidden with a styled <div>4. The hard part is to place the <div>s in appropriate places. To do that you will need to do some relatively simple parsing. We'll give you some help with that.